Renaame cli & web package names; move plugin code to lib/plugin.
