+++ packages/plugin-trpc/cli/src/templates/base/trpc.ts.ejs
@@ -1,4 +1,4 @@
-import { initTRPC } from 'webstone-plugin-web-trpc';
+import { initTRPC } from 'webstone-plugin-trpc-web';
 const t = initTRPC.create();
 
 export const middleware = t.middleware;
